+++ etna/transforms/math/statistics.py
@@ -136,6 +136,14 @@ class MeanTransform(WindowStatisticsTransform):
 
     .. math::
        MeanTransform(x_t) = \\sum_{i=1}^{window}{x_{t - i}\\cdot\\alpha^{i - 1}}
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
     """
 
     def __init__(
@@ -235,6 +243,14 @@ class MeanTransform(WindowStatisticsTransform):
 class StdTransform(WindowStatisticsTransform):
     """StdTransform computes std value for given window.
 
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+
     Notes
     -----
     Note that ``pd.Series([1]).std()`` is ``np.nan``.
@@ -293,7 +309,16 @@ class StdTransform(WindowStatisticsTransform):
 
 
 class QuantileTransform(WindowStatisticsTransform):
-    """QuantileTransform computes quantile value for given window."""
+    """QuantileTransform computes quantile value for given window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,
@@ -367,7 +392,16 @@ class QuantileTransform(WindowStatisticsTransform):
 
 
 class MinTransform(WindowStatisticsTransform):
-    """MinTransform computes min value for given window."""
+    """MinTransform computes min value for given window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,
@@ -418,7 +452,16 @@ class MinTransform(WindowStatisticsTransform):
 
 
 class MaxTransform(WindowStatisticsTransform):
-    """MaxTransform computes max value for given window."""
+    """MaxTransform computes max value for given window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,
@@ -469,7 +512,16 @@ class MaxTransform(WindowStatisticsTransform):
 
 
 class MedianTransform(WindowStatisticsTransform):
-    """MedianTransform computes median value for given window."""
+    """MedianTransform computes median value for given window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,
@@ -520,7 +572,16 @@ class MedianTransform(WindowStatisticsTransform):
 
 
 class MADTransform(WindowStatisticsTransform):
-    """MADTransform computes Mean Absolute Deviation over the window."""
+    """MADTransform computes Mean Absolute Deviation over the window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,
@@ -577,7 +638,16 @@ class MADTransform(WindowStatisticsTransform):
 
 
 class MinMaxDifferenceTransform(WindowStatisticsTransform):
-    """MinMaxDifferenceTransform computes difference between max and min values for given window."""
+    """MinMaxDifferenceTransform computes difference between max and min values for given window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,
@@ -630,7 +700,16 @@ class MinMaxDifferenceTransform(WindowStatisticsTransform):
 
 
 class SumTransform(WindowStatisticsTransform):
-    """SumTransform computes sum of values over given window."""
+    """SumTransform computes sum of values over given window.
+
+    Warning
+    -------
+    This transform, applied to non-regressor column, generates non-regressor column.
+    Apply it to regressor columns to get regressor columns too.
+    In the majority of cases you need to generate regressor to use them in the future.
+
+    For example, apply this transform to target lags, not to target directly.
+    """
 
     def __init__(
         self,